+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * NNLS merging now also accepts a list of spatial moments to preserve (requires explicitly conserving 1-st order spatial moments, otherwise
     the corresponding coordinates are set to 0)
 * 1D example added for NNLS merging (`simulations/1D/couette_varweight_nnls.jl`)
+* NNLS merging speed-ups
+* New weighted median with interpolation function for octree merging
 
 ## v0.7.1
 * More use of Muladd in a few places